requirements.txt
tensorflow
pandas
numpy
keras
matplotlib
nltk
scikit-learn
minio
joblib
networkx
apache-airflow-providers-apache-spark